Run Importer: Import Cleanup: Delete Failed changes
Exclude Failed for System Error in delete before 3 days.
This allows failed for System Error to be requeued.

+++ limelight_run_importer/src/main/java/org/yeastrc/limelight/limelight_run_importer/import_files_remove_success_failed_except_last_2_main_and_searcher/ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main.java
@@ -1,20 +1,28 @@
 package org.yeastrc.limelight.limelight_run_importer.import_files_remove_success_failed_except_last_2_main_and_searcher;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.yeastrc.limelight.limelight_importer_runimporter_shared.db.ImportRunImporterDBConnectionFactory;
 import org.yeastrc.limelight.limelight_run_importer.delete_directory_and_contents.DeleteDirectoryAndContents;
+import org.yeastrc.limelight.limelight_run_importer.import_files_remove_success_failed_except_last_2_main_and_searcher.ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers.ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers__getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultItem;
 import org.yeastrc.limelight.limelight_shared.config_system_table_common_access.ConfigSystemTableGetValueCommon;
 import org.yeastrc.limelight.limelight_shared.config_system_table_common_access.ConfigSystemsKeysSharedConstants;
 import org.yeastrc.limelight.limelight_shared.config_system_table_common_access.ConfigSystemsValuesSharedConstants;
 import org.yeastrc.limelight.limelight_shared.file_import_limelight_xml_scans.constants.FileUploadCommonConstants;
+import org.yeastrc.limelight.limelight_shared.file_import_limelight_xml_scans.enum_classes.FileImportRunSubStatus;
 import org.yeastrc.limelight.limelight_shared.file_import_limelight_xml_scans.enum_classes.FileImportStatus;
 import org.yeastrc.limelight.limelight_shared.file_import_limelight_xml_scans.utils.Limelight_XML_ImporterWrkDirAndSbDrsCmmn;
 
 /**
- * 
+ * Remove excludes Failed for System Error.
+ * Keep Failed for System Error until the 3 day delete to support requeue after the System Error has been addressed.
  *
  */
 public class ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main {
@@ -110,7 +118,7 @@ public class ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main {
 	 */
 	private void internal_ProcessMain(ImportRunImporterDBConnectionFactory importRunImporterDBConnectionFactory) throws Exception {
 		
-		{  //  Delete All Status Success Except Last 2
+		{  //  Delete All Status 'Success' Except Last 2
 			
 			List<Integer> fileImportTrackingIddList_To_Delete_Directories =
 					ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers.getSingletonInstance()
@@ -127,21 +135,79 @@ public class ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main {
 			delete_Directories_For_trackingIdList(fileImportTrackingIddList_To_Delete_Directories);
 		}
 
-		if ( keepRunning ) {  //  Delete All Status Fail Except Last 2
+		if ( keepRunning ) {  //  Delete All Status 'Fail' Except Last 2, excluding Fail for "System Error"
 			
-			List<Integer> fileImportTrackingIddList_To_Delete_Directories =
-					ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers.getSingletonInstance()
-					.getAll_TrackingId_For_Status( FileImportStatus.FAILED, importRunImporterDBConnectionFactory);
+			List<Integer> fileImportTrackingIddList_To_Delete_Directories = null;
+			
+			{
+				List<Integer> fileImportTrackingIddList_ForFailedStatus =
+						ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers.getSingletonInstance()
+						.getAll_TrackingId_For_Status( FileImportStatus.FAILED, importRunImporterDBConnectionFactory);
 
-			//  Delete last 2 entries if exist
-			if ( fileImportTrackingIddList_To_Delete_Directories.size() >= 2 ) {
-				fileImportTrackingIddList_To_Delete_Directories.remove( fileImportTrackingIddList_To_Delete_Directories.size() - 1 );
-			}
-			if ( fileImportTrackingIddList_To_Delete_Directories.size() >= 1 ) {
-				fileImportTrackingIddList_To_Delete_Directories.remove( fileImportTrackingIddList_To_Delete_Directories.size() - 1 );
+				if ( ! fileImportTrackingIddList_ForFailedStatus.isEmpty() ) {
+
+					//  Remove all fileImportTrackingId where last run is System Error
+
+					Map<Integer, ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers__getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultItem> trackingId_TrackingRunId_SubStatus_Map_Key_TrackingId = 
+							new HashMap<>( fileImportTrackingIddList_ForFailedStatus.size() );
+
+					List<ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers__getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultItem> getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultList = 
+							ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers.getSingletonInstance()
+							.getAll_TrackingId_TrackingRunId_SubStatus_For_Status( FileImportStatus.FAILED, importRunImporterDBConnectionFactory);
+
+					for ( ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers__getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultItem trackingId_TrackingRunId_SubStatus_Entry : getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultList ) {
+						
+						ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers__getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultItem entryFromMap = 
+								trackingId_TrackingRunId_SubStatus_Map_Key_TrackingId.get( trackingId_TrackingRunId_SubStatus_Entry.trackingId );
+						if ( entryFromMap == null ) {
+							trackingId_TrackingRunId_SubStatus_Map_Key_TrackingId.put( trackingId_TrackingRunId_SubStatus_Entry.trackingId, trackingId_TrackingRunId_SubStatus_Entry );
+						} else {
+							if ( trackingId_TrackingRunId_SubStatus_Entry.trackingRunId > entryFromMap.trackingRunId ) {
+								//   New entry has larger run id (later run) so replace entry in map with new entry
+								trackingId_TrackingRunId_SubStatus_Map_Key_TrackingId.put( trackingId_TrackingRunId_SubStatus_Entry.trackingId, trackingId_TrackingRunId_SubStatus_Entry );
+							}
+						}
+					}
+					
+					fileImportTrackingIddList_To_Delete_Directories = new ArrayList<>( fileImportTrackingIddList_ForFailedStatus.size() );
+					
+					for ( Integer fileImportTrackingId_FailedStatus : fileImportTrackingIddList_ForFailedStatus ) {
+						
+						boolean deleteDirFor__fileImportTrackingId_FailedStatus = true;
+						
+						{
+							ImportFiles_Remove_SuccessFailed_ExceptLastTwo_Main_Searchers__getAll_TrackingId_TrackingRunId_SubStatus_For_Status_ResultItem entryFromMap = 
+									trackingId_TrackingRunId_SubStatus_Map_Key_TrackingId.get( fileImportTrackingId_FailedStatus );
+
+							if ( entryFromMap != null && entryFromMap.trackingRun_SubStatusId == FileImportRunSubStatus.SYSTEM_ERROR.value() ) {
+
+								deleteDirFor__fileImportTrackingId_FailedStatus = false;  // NOT delete System Error
+							}
+						}
+						
+						if ( deleteDirFor__fileImportTrackingId_FailedStatus ) {
+							fileImportTrackingIddList_To_Delete_Directories.add( fileImportTrackingId_FailedStatus );
+						}
+					}
+				}
 			}
 			
-			delete_Directories_For_trackingIdList(fileImportTrackingIddList_To_Delete_Directories);
+			if ( fileImportTrackingIddList_To_Delete_Directories != null ) {
+				
+				//  Sort in ascending order
+				
+				Collections.sort( fileImportTrackingIddList_To_Delete_Directories );
+
+				//  Delete last 2 entries if exist
+				if ( fileImportTrackingIddList_To_Delete_Directories.size() >= 2 ) {
+					fileImportTrackingIddList_To_Delete_Directories.remove( fileImportTrackingIddList_To_Delete_Directories.size() - 1 );
+				}
+				if ( fileImportTrackingIddList_To_Delete_Directories.size() >= 1 ) {
+					fileImportTrackingIddList_To_Delete_Directories.remove( fileImportTrackingIddList_To_Delete_Directories.size() - 1 );
+				}
+
+				delete_Directories_For_trackingIdList(fileImportTrackingIddList_To_Delete_Directories);
+			}
 		}
 	}
